+++ examples/ExampleFeatures1.hs
@@ -4,7 +4,6 @@ Description : Demostrates how to define features using Hasklepias
 Copyright   : (c) NoviSci, Inc 2020
 License     : BSD3
 Maintainer  : bsaul@novisci.com
-Stability   : experimental
 -}
 
 {-# LANGUAGE OverloadedStrings #-}