+++ examples/ExampleFeatures3.hs
@@ -1,0 +1,58 @@
+{-|
+Module      : ExampleFeatures3
+Description : Demostrates how to define features using Hasklepias
+Copyright   : (c) NoviSci, Inc 2020
+License     : BSD3
+Maintainer  : bsaul@novisci.com
+-}
+
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE AllowAmbiguousTypes #-}
+{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE MonoLocalBinds #-}
+module ExampleFeatures3(
+    exampleFeatures3Spec
+) where
+
+import Hasklepias
+import ExampleEvents
+import Test.Hspec
+import Data.Text(pack, Text)
+import Data.Maybe ( fromMaybe )
+import Data.Set(member)
+
+
+examplePairComparison :: (IntervalAlgebraic Interval a
+              , IntervalSizeable a b) =>
+               FeatureDefinition (Interval a) a (Bool, Maybe a)
+examplePairComparison = defineFEF Unknown
+    ( \i es -> 
+       (  (\x ->                            
+                ( isNotEmpty x              -- are there any?
+                , fmap begin (safeLast x))) -- if exists, keep the begin of the last "c1" interval
+
+          . intervals                       -- get the intervals of any "c1" events
+          . map fst                         
+
+          . filter                          -- filter this list of pairs to cases 
+            (\pr -> fst pr `concur`         -- where "c1" event concurs with +/- 3
+                    expand 3 3 (snd pr) )   -- of any "c2" event
+
+          . uncurry allPairs                -- form a list of pairs where first element
+          . splitByConcepts ["c1"] ["c2"]   -- has "c1" events and second has "c2" events
+                                           
+          . filterConcur i )                -- filter to concurring with followup interval
+        es
+    )
+
+flwup :: Feature (Interval Int)
+flwup = featureR $ beginerval 50 0
+
+exampleFeatures3Spec :: Spec
+exampleFeatures3Spec = do 
+
+    it "examplePairComparison"  $
+        applyFEF examplePairComparison flwup exampleEvents4 `shouldBe`
+        featureR (True, Just 16)
+