+++ examples/Main.hs
@@ -9,8 +9,10 @@ import Hasklepias
 import ExampleEvents
 import ExampleFeatures1
 import ExampleFeatures2
+import ExampleFeatures3
 import Test.Hspec ( hspec )
 
 main :: IO ()
 main = hspec $ do exampleFeatures1Spec
-                  exampleFeatures2Spec 
+                  exampleFeatures2Spec
+                  exampleFeatures3Spec